Encapsulate Intent resolving logic into a new class IntentResolver
